Add runtime compatbility check
Conditions:
- Major version must match bindings exactly
- Minor version must be greater than or equal to bindings version

In practice, things *may* work even if the bindings and runtime versions
do not match, but proving this requires carefully checking for API/ABI
breaks in all possible workflows.

For example: the image decoding workflow in this project breaks when
attempting to use ffmpeg 6.1 with ffmpeg 5.1 bindings due to changes to
API changes in ffmpeg 6.0.